Strip unused files from the build
This patch strips out a significant portion of the uncompiled and
unused code. Lean, mean, low-level-machine here we come!

Qemu is feature rich, even this old base that xqemu is based on. For
general Xbox emulation, many of the features of Qemu seem to be pretty
overkill--this means the code is bigger, slower, and harder to
maintain. Since we don't have any real near-term plans to rebase or
upstream to Qemu proper, the goal of this patch is to streamline the
code base as much as possible to make it more maintainable, managable,
and less overwhelming to new developers. Then we can refactor, improve,
and expand the code base.