Fix TSC/ACPI PM timer
This patch ports over the following patches from xqemu1:
- e3d643e Fix the TSC at 733 MHz
- 81dd218 Make the ACPI PM timer more hardware accurate

I've also added a new macro, "XBOX", to identify some non-configurable
changes from QEMU mainline. With a bit of restructuring this macro might
not be needed, but this is the path of least resistance right now and
will help with future rebasing/merging efforts to identify the code that
has been modified for Xbox emulation.