apu: Connect VP output directly into system audio
This patch gets *very* basic audio working in XQEMU by feeding the
output of the VP directly into system audio through the QEMU sound
architecture.

Note: the audio coming out of the VP is not fully mixed, and this patch
mixes all voices down to a single channel to keep things simple for now.
Currently, only buffers are played, streamed audio is not yet supported.
This may cause some games to hang. Furthermore, the DSP is not involved
here yet, so in addition to poor mixing, some audio may be horribly
broken.